Refactor: introduce indirection between HTTP server and Trillian calls
The Trillian Client class that we've used thus far can't take us any further; it checks the consistency & inclusion proofs, but swallows them and thus doesn't allow the personality access to them to pass them on to its clients. The new Client class will replace the LogClient we were using by duplicating some of its internals, but for the sake of this refactor we are just wrapping it for now.

There is a slight functional change here; previously we didn't return a HTTP status failure if AddLeaf failed. This will now return an error.

Change-Id: I9f2f0f69aee8d4feab800ec109ab00665fc3bacd