Implemented get-manifest-entries
This version simply gets a single leaf. We'll almost certainly want to make this get by range, and/or fetch the manifest entries with a given hash. However, this should provide the bare minimum to complete the v0.0.1 demo personality; data can be written to the log and fetched back out, and the client can ensure they are seeing a consistent append-only evolution of the log.

Change-Id: I730f3ab458e15148fe1c165a6b3f88bb078bc23d