Generate a verifiable map from sumdb auditor DB
This generates a map where the keys are the module+version and a component
to indicate whether it is the repo hash or the go.mod hash, and the value
is the hash value string as found in the go.sum file.

The verifier can (inefficiently) check that all values in a go.sum file
are committed to in the verifiable map.