Minor improvements to SumDB auditor
Added documentation and better console logging when running. None of this is critical, but it's all stuff that is clear after being away from this code for a while.

The reason for revisiting this code is that it will be changed to persist the Golden Checkpoint between runs. This improves the security for the stragglers, and it will give the map demo that builds on this access to a Log Checkpoint to write in its Map Checkpoint to allow traceability.