Minor improvements to SumDB auditor (#170)
Added documentation and better console logging when running. None of this is critical, but it's all stuff that is clear after being away from this code for a while. Also made the error handling stricter in case the DB schema is corrupt when reading the latest tree head.

The reason for revisiting this code is that it will be changed to persist the Golden Checkpoint between runs. This improves the security for the stragglers, and it will give the map demo that builds on this access to a Log Checkpoint to write in its Map Checkpoint to allow traceability.