Propagate the log checkpoint information into a new map revisions table
This provides enough information for another party to confirm they have the same view of the source log (using logroot consistency checks) and then to recreate the map from their mirror using the count.

Any writes that fail may have some of their tiles written, but won't get the entry in the revisions table. This is as close as we can get to a write transaction using Beam.