Created a mirror service which continually runs
The clone.go tool is intended to be a one-shot tool, and a gentle introduction/demo. This tool is intended to be run as a service which can continually keep a database up to date.

Future work might be to support different databases (e.g. MySQL starts to look reasonable for something like this), or to have some documentation on how to set this up as a Linux service.