Support for building logs of versions for a module (#224)
If the relevant flag is specified when running the pipeline then this creates a log per module, containing the versions that SumDB reports for it. The root hash of each of these logs is committed to by including it in the map. The log leaf data (i.e. the version strings) are available in a new `logs` table.

A future PR will add a tool to verifiably list the versions for a module.

There is some refactoring to extract pipeline code from the main file runner in this PR too.