Added support for client/server SQL databases (#228)
Having the data in a client/server database opens the way for running a mirror in GCP which can be used to build a verifiable map in dataflow. This is infeasible when the database is a local file.

This has been tested with MariaDB 10.5.8.

N.B. This contains a breaking DB schema change. TEXT is not a great choice because at least MariaDB will group by this case insensitively, which surprisingly caused "duplicates" to be identified. Using a binary representation for these fields avoids such false positives.