Tool to list verified versions for a module (#225)
This builds on #224 by adding a tool to make use of the log of versions. It pulls the list of versions, constructs the log for it, and checks that the map commits to this log. By default it prints only release versions, but the --all flag can be used to list ephemeral versions too (e.g. v0.0.0-20191025170756-f12f9191a1a4).