Personality can be deployed in a single command
The personality is now deployed at the same time as Trillian if the docker command in the README is followed.

This requires an extra table of storage for the tree details so that the container can be restarted. The service will now create the tree if it hasn't done so already. This saves quite a lot of human-driven configuration from the start of the demo, which massively streamlines it (and mitigates the risk of anything going wrong from incorrect configuration). Having fewer terminals also lowers the cognitive cost for the audience.