Construct a map of logs from FT log
The map is keyed by device ID, and the value is a log which contains the firmware revisions for that device. The log is stored in the database as a flat JSON list of leaf values, and will need to be rehydrated (i.e. the merkle tree recalculated) each time it is used. This puts a limit on how big the sub-logs can be, but that limit is much higher than the cardinalities we'll use for demos.

This requires that the `mysql` docker container has a port exposed outside of the docker network so that the Beam pipeline can access the DB to source the log data.

This pipeline is something of a strawman. By starting with a map of logs, this demonstrates the most complicated map construction and thus all the raw material should be here to allow developers with less experience with Beam to mould this to take whatever shape of Map we decide is required for a convincing demo.