Annotation API (#241)
API for adding annotations to the log

These annotations will be stored in the main log, alongside the firmware metadata. This involved changing FirmwareStatement to support multiple different logical statement types in a uniform schema. No tooling is provided in this PR for the malware annotator to publish.

This is a breaking change, and anyone with a database in the old format will need to start afresh.