Updated Trillian and made required changes
This is a big one as the interactive/DB-backed Map has been removed from Trillian, which meant that the `registers` and `etherslurp` examples were left without the library they were demonstrating. They could be fixed up to use the new experimental batchmap if this is a direction we want to go in, but I suggest we drop them and re-add them, or we risk falling further behind Trillian while this work is done.

`HStar2` was also removed, so some verification code was updated to use `HStar3` instead.