Aggregate the FW with the annotations (#275)
The pipeline now creates an entry for each piece of firmware in the log with an aggregated version of the malware annotation. The aggregation rule is that it is good, and will switch to bad if there exists any malware annotation saying it is bad. We may wish to revise this to be the result of the most recent annotation, but this approach has the benefit of simplicity. And it seems like an edge-case that a release would be classified as malware, and then the scanner would change its mind.

The key in the map is derived from the Log Index of the original FW metadata, and the value is a commitment to a JSON-ecoded `AggregatedFirmware`.
The map tiles and aggregated firmware are stored in a DB, but there is not currently any server for returning this data to clients.