Updated to latest version of trillian
All the changes are fallout from either direct changes to trillian API, or because the update to trillian updated the proto API, and that now has a bunch of stuff which is deprecated and thus blocking due to our usage of the linter.

This fixes the nil signature problem seen because we're running FT against a non-pinned version of Trillian images. The fix is very indirect; it's the trillian client that removes this check.