Foundations for map server (#297)
This currently implements only the map checkpoint method.

This changes the embedded LogCheckpoint in the Map Checkpoint from an unmarshalled type to being raw bytes. This will be important for clients to validate the log checkpoint signature eventually, and it will form a packet of evidence should this identify a split view.