Remove unused nested field
Keeping this forces the map server to be able to return the raw log entries. This could be done, but requires either a connection to another DB, or duplicating data. Furthermore, the client already has this data and was just asserting equality. Given that we've already checked log consistency, we're already convinced we have the same view of the log data.