Persist the monitor state to avoid repeated work
Without this, the monitor starts from the beginning of the log on each run. This is wasted effort in the best-case, but can actually cause a lot of spam annotations to the log in the worst case.

The state file can always be removed locally if the monitor code is updated and needs to process the whole log again.