Tool that bulk downloads CT logs, respecting backoffs (#338)
The tool clones leaves from the log as quickly as possible to a MySQL instance provisioned by the user. This is the foundation for doing local processing of all data, such as building a verifiable map.

This is optimized very heavily towards downloading a large amount of data quickly, while respecting the log if it asks the client to back off. Missing features:
 * downloading/persisting checkpoints & verifying the leaf data
 * terminating cleanly when the end of the log is reached

The design ensures that leaves are never written without any gaps in order to simplify the local state required.