Move DB out of internal for use in tools that need to read the content (#360)
The DB is the API for the clone tool. I considered separating this into write/read, and keeping the write interface as internal. But that seemed like overkill. We can revisit this if desired.

The reading API is needed for the CT map prototype, which is the motivation for this change.