Support writing & reading checkpoints in DB (#399)
Checkpoints are stored with the compact range, which allows the log contents to be verified incrementally.

Verified checkpoints are now stored by the `ctverify` tool. Upon verifying a new checkpoint, the latest verified checkpoint will be loaded and only the leaves added since that checkpoint will directly be checked.

This is a performance improvement for verification, and it also provides tooling that reads from the log (e.g. map building code) to find the log checkpoints. This code can either trust the local DB hasn't been modified since verification, or it can re-verify all leaves against the checkpoint if desired. In the case of the map, the log checkpoint used will be embedded within the map checkpoint in order to complete an audit trail.