Beam pipeline to create a map of logs, verifiably indexing a log (#361)
The map has keys are all of the raw domain names in the SAN. Certs without any DNS names in the SAN are dropped.

For each DNS name, the value in the map is a log of all indices at which this domain has a cert for it. The actual value committed to by the map is the root of this log.

At the moment this generates a single huge file which contains all domains and the list of indices. Clients could download this and generate the logs locally for the domains they are interested in, and then check the log root hash is in the map.

This is a proof of concept, and does not represent how this would likely be served in a real production setup.