Witness client: Update returns latest Checkpoint (#435)
The witness client returns the latest checkpoint when it is available so that feeders don't need to make another call to GetCheckpoint.

It also includes a small fix on the witness to not add a trailing newline to the checkpoint when we return StatusConflict. This avoids stripping characters on the client to avoid malformed notes.