Use locking in reporter to avoid anomolies
This was using a lock-free mechanism before, which worked but isn't a good practice.

This change was initially created to avoid seeing > 100% work in any reporting window. It seems to reduce that but not completely avoid it. The reason for that is that work isn't cleanly divided into reporting chunks and can overlap reporting periods, but only be counted in one of them.