SumDB clone tool that downloads the DB quickly  (#453)
This is consistent with the ctclone tool and uses much of the same machinery to get the job done. There is a little additional complexity here where we need to always be downloading a tile per worker to realize the parallelism dream. This means that when the previous download finished mid-tile, we need to do a little neatening up before we can invoke the Bulk command. Otherwise, this drops in cleanly if I say so myself.