Extracted common configs into separate package (#497)
In the process of attempting to extract a utility to push the latest checkpoint from a witness to a distributor, I found myself once again copying and pasting these definitions. Having them be common allows fragments describing these entities to be easily copied into config files.

The exception to this clean extraction is the Log.ID which is not really a property of the log. I've added a TODO to come back to this.