Added default value for log identifier (#510)
This is determinstic given an origin string and public key, and is guaranteed to contain no characters that could abuse filesystem or URL path handling. I've removed the ID from the config files so this is more likely to be the default, though this can still be overridden if users really want to.