Added trigger for auto-merge completing (#520)
This doesn't trigger the main branch push trigger for some reason. It could be that there's no harm in leaving this uncommented and it just won't trigger if such a workflow doesn't exist, but I'm erring on the side of caution.

Also given the workflow a name, and allow it to be manually triggered in case that's needed.