Added cloudbuild script that builds docker images (#521)
For now this is starting small and we'll build up to all of the docker images if that's useful. These images are picked based on being the ones I run on a very old Raspberry Pi, on which they take an age to build.

We'll trigger this build script manually for now, but we can make this nightly or triggered by tags in the future. It's probably too heavy to put in the PR cloudbuild, hence keeping it separate.