Make the docker images we build as small as possible (#522)
The previous images were just shy of 1GB each. The new ones are under 20MB. I'm gonna call that a win.

After this is in and images pushed to gcr.io, I'll update the docker-compose files to use these images. This should bring down the time needed to deploy a sumdb witness by an order of magnitude.