Unikernel and single executable feeder & witness
This runs the omniwitness feeder & witness, but purely in-memory.
Until this is power cycled, this performs most of the same work that
the docker omniwitness does. The distributors aren't currently run,
this will be added in future PRs.

`make CROSS_COMPILE=arm-none-eabi- imx` builds the imx file

Note that this disables TLS and performs manual DNS in order to support
network operations on a unikernel running on the usbarmory.

Note that this uses build tags to keep the rest of the tree free from
tamago-specific build concerns. This means `go build ./...` still works
 at the root, for example.