Improve feeder interfaces and timeout handling
All feeders now take a http client used for all http connections. Furthermore, they all take the interface definition for a witness instead of the http client implementation. This allows different implementations to be passed in

All clients configure the http client with a timeout. Previously, it doesn't look like the timeout was actually being used in many cases.