Abstraction layer for witness persistence (#600)
This allows the same witness to be used with different storage layers other than SQL. In particular, this allows a lightweight version that doesn't use SQL or CGO to be implemented that can be compiled as a unikernel using tamago (see #601 for motivation). At the moment, the provided alternative is purely inmemory. This is great for testing, and works as a proof of concept, but we'll need something that writes to persistent storage for any real deployment.

Changed tests over to use inmemory persistence.