Make SumDB Feeder configured similarly to the others
This feeder was implemented before the others, and at the time it seemed like a good idea to have it take the URL from the key name. This now takes the URL explicitly, which makes it consistent with the others. The config for this feeder in the omniwitness now lives in a config file.

Note that this is based on #601. We could sequence them the other way
around, but I've done it this way as #601 has already been through one
round of review.