Make SumDB Feeder configured similarly to the others (#606)
This feeder was implemented before the others, and at the time it seemed like a good idea to have it take the URL from the key name. This now takes the URL explicitly, which makes it consistent with the others. The config for this feeder in the omniwitness now lives in a config file.