Distributor incoming files have unique names
Previously we were creating the filename based on the unsigned checkpoint body, which meant that there would be a collision when two different witnesses attempted to distribute the same checkpoint. Using the witness name as a salt means that this collision will be avoided, which should allow for better throughput, latency, and scaling for distributing checkpoints for the same log when there are multiple witnesses.