dry_run mode checks all log directories (#621)
The dry_run check terminates after seeing that one incoming directory is fine. This allows a later incoming directory to be broken, but the automerge check to pass.

This is happening at the moment because of the Pixel BT changing ID. The PR is approved because one of the other log directories is fine, but then none of the checkpoints can be integrated because the old pixel one is uknown.