Build the monolithic omniwitness as Docker image
This may replace the docker-compose version of the omniwitness, at least as our primary recommendation of how to run this. The primary benefit is that it's much simpler; there's only one executable to launch and the only config it takes are those that are user-specific. All of the details of the logs and github distributors are baked into the binary. This should make it far easier to ensure everyone is running the same configuration of the omniwitness as logs are added, etc.

There is explicitly no plan to prevent users running these components flexibly, this is simply a case of having the omniwitness very opinionated to encourage simple adoption.