Override http.Client used by distributors
This was implicitly using the default client, which doesn't work on the USB Armory as it requires a custom client. This change sets the client via a context, which is the way the oauth2 libraries like to do it. This could also have been implemented by putting the client in the context at a higher level instead of passing it around, but the explicit argument passing is clearer and more consistent with how we've implemented the rest of this code.