Upgrade lint to more recent version and fixed errors (#655)
This started off as trying to make the lint configuration consistent with other repos, and removing skip instruction that no longer matched anything. In doing so, it turned out that we were using an old version, and updating that caused a ream of lint warnings. So I've fixed them.

This also brought to light some weirdness where we install multiple versions of go (because we install tamago). I've made it so that:
a) the same version of both go and tamago is installed
b) tamago doesn't blat go in the path; invoking go will do normal go, not tamago