Clone tool for serverless logs
This clones a serverless transparency log made available over HTTP(s) into a local database, and verifies that the downloaded data matches the checkpoint.

Once the data is present locally, it can be analyzed by tooling in any language that can read the data from SQL. This allows quick searching or verification code to be written by interested parties, without needing to integrate with Go or understand Merkle proofs.