Updated all go versions to minimum of 1.17 (#669)
This is what trillian is using now. It seems unlikely someone would need higher levels of backwards compatibility in the examples repo.

`go mod tidy` caused a lot of changes.
https://go.dev/doc/modules/gomod-ref notes that `The go.mod file
includes an explicit require directive for each module that provides any
package transitively imported by a package or test in the main module`
which changed with 1.17.