Fix Rekor feeder to handle sharded logs
This updates the feeder to specify the shard ID when fetching a
consistency proof, using the ID from the config.

I also needed to update fetching the latest checkpoint, searching for
the checkpoint that matches the shard ID from the config.

Also added the new shard to the configs, and updated the Origin to
match what's now returned by Rekor.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>