Allow inclusion proofs to be looked up by node hash.
This adds a command-line argument to the client command-line
tool that allows the entry to be looked up by merkle node hash
instead by a file containing the entry itself.

Note that the hash here is the rfc6962 node hash and not just the
sha256 hash of the node's contents. The hashed data is a single
0 byte followed by the node's contents.