Optionally output log inclusion proof in JSON.
For some scenarios, it is useful to independently check a log
inclusion proof. The serverless "client inclusion" command does
provide the information necessary to verify the proof, such as
the node index, the tree size, the tree root hash, and node hashes.
But these are not provided in an easy-to-consume format. The
hashes can be written to a file, but they are useless without the
node index, tree size, and root hash. The index, tree size and root
hash have to be scraped from the log output, which requires messy
text parsing.

This change adds a new option to the "client inclusion" command
to write all of this information to a file in JSON format.
This gives us a self-contained proof file that can be used to
verify log inclusion.