Improve clone tool docker integration (#774)
This uses explicit Go versions for the builder, and uses alpine for both the builder image and the release image, which is much smaller. This uses `docker compose` instead of `docker-compose`, which moves us to newer code written in Go instead of Python.

There is now a docker-compose for sumdb that brings up a mariadb image and the sumdb cloner that will clone into it.