Use MariaDB instead of sqlite in tests
This uses docker to bring up a temporary database server, and each test creates a new database within it to prevent data collisions.

GCB is a bit of a problem in that it has some custom configuration required on the docker host. This problem is worked around by having a build tag for cloudbuild; when this is provided then the `ConfigureHost` function is defined to do this configuration, otherwise it does nothing.