Allow witnesses to be configured at runtime (#820)
The witnesses that a distributor accepts should not be statically compiled into the code as we don't have a strong opinion on this set. The logs could also be made configurable at runtime in the future, but for now keeping these the same as the omniwitness makes configuration simpler and doesn't alienate any users.

Also updated ports to avoid collisions with witness.