Added tool for creating tile-based logs from cloned log
This tool allows a file-based mirror of any cloneable log to be created. This required a change to Integrate to remove the requirement for a prior checkpoint. This has been replaced with the fromSize. In the case of mirrors, we can't create and sign a valid checkpooint of size 0, which is what the integrate command had assumed was possible from its previous usage.

TODO: write up instructions for setting up a filesystem that can handle this without running out of space / inodes.