Improve the StreamLeaves API and implementation (#842)
This is a long-running operation and now respects context cancellation.

Also changed the two returns channels to combine them into a single channel. This is far easier to understand and implement bug-free code with. Previously the err channel wasn't closed, which wasn't clean. Now everything is closed when StreamLeaves returns.