Added tool for creating tile-based logs from cloned log (#852)
This tool allows a file-based mirror of any cloneable log to be created. For users with a "normal" filesystem setup, following the setup instructions in the README is strongly encouraged before running this on a large log!

This required a change to Integrate to remove the requirement for a prior checkpoint. This has been replaced with the fromSize. In the case of mirrors, we can't create and sign a valid checkpooint of size 0, which is what the integrate command had assumed was possible from its previous usage.