Added verifier for sumdb claims (#978)
This checks all of the cloned entries from the sumdb log and checks for any entries with the same module and version, but different hashes.

Any conflicts will be logged to the error stream. The final conflict will be returned up as an error.

This has been written as a reference verifier, and demonstrates best/required practices for verifiers, such as ensuring the data verified is committed to by the checkpoint.