Verifiable Index skeleton (#1293)
This is a shared place for those of us that have been discussing these
ideas to collaborate and nudge things forward in between main projects.

There's still no code here to build the map, but this shows how the map function will be applied and data marshalled. Still need to update the files on disk as the WAL is updated.

Current thinking is that the WAL should be changed to not have a single file, but be given a directory. This will allow it to move the old file to a backup, and then open the original file in append-only, and write out the entries that were restored from, and then keep appending new entries, flushing periodically.