[VIndex] Rewrite of the WAL code (#1327)
The old code had a number of issues with correctness and performance
(e.g. reading the whole file to get the last line). This rewrite moves
things into a more useful direction.

Main changes:
 - WAL fails to init for broken files: this means that human
   intervention is currently required in these cases. Future work may
   try to fix some easy cases, but for now, a successful init means that
   the file can be appended to.
- WAL is now written to disk by the IndexBuilder
- WAL reader is implemented and confirmed working via tests
- Index returned from init is now clear: before 0 could be returned for
  the empty file, or a file where 0 had been written correctly.

This PR leaves the repo in a sensible place on which to hook in a map
implementation (in-memory at first), which can read the WAL at startup
to populate itself.