End-to-end index built via a command line tool
This version is the MVP to demonstrate creating an index out of a clone
of the SumDB. Each release with a plain semver (e.g. vX.Y.Z, no suffix)
is logged into the map. The map key is the hash of the module name. The
map value is the list of all uint64 indices at which a release for this
module can be found in the origin log.

An HTTP server is launched, which has a simple form that takes a map key
and looks up the corresponding indices.

Under the hood this is constructing a Merkle tree, though it isn't used
for Lookup yet.