Improve Update
cmd now periodically updates

Added basic locking over data structures for update and lookup. The
vindex takes a while to create the first time, and this blocks serving
while it's happening. There's scope to improve this, but better than
having no locking.

Improved update to keep a note of which keys it has updated, and only
update those keys in the vindex update. This doesn't improve the first
init speed, but makes it much faster for incremental updates.