End-to-end index built via a command line tool (#1328)
* End-to-end index built via a command line tool

This version is the MVP to demonstrate creating an index out of a clone
of the SumDB. Each release with a plain semver (e.g. vX.Y.Z, no suffix)
is logged into the map. The map key is the hash of the module name. The
map value is the list of all uint64 indices at which a release for this
module can be found in the origin log.

An HTTP server is launched, which has a simple form that takes a map key
and looks up the corresponding indices.

Under the hood this is constructing a Merkle tree, though it isn't used
for Lookup yet.

* Filippo comments addressed

* Check errors; make linter happy

* Process vindex as a batch

Doing it inline led to a lot of duplicate work, and updating the index
is the slowest part of the processing.

This brought the time down from several minutes to ~30s.

```
go run ./experimental/vindex/cmd --logDSN="sumdb:letmein@tcp(127.0.0.1:33006)/sumdb"  --walPath ~/sumdb.wal
I0610 16:10:24.470289  570448 main.go:93] HTTP server listening on :8088
I0610 16:10:24.470721  570448 map.go:110] Building map to log size of 38787558
I0610 16:10:56.099454  570448 map.go:204] buildMap: total=31.628709524s (wal=8.081146649s, vindex=23.547562536s)
I0610 16:10:56.099483  570448 main.go:85] Verifiable Index built
```

* Improve Update

cmd now periodically updates

Added basic locking over data structures for update and lookup. The
vindex takes a while to create the first time, and this blocks serving
while it's happening. There's scope to improve this, but better than
having no locking.

Improved update to keep a note of which keys it has updated, and only
update those keys in the vindex update. This doesn't improve the first
init speed, but makes it much faster for incremental updates.

* Update docs with latest implementation, how to run, and roadmap

* Address Al's comments

Biggest change wraps the mapFn invocation in a func so we can catch any panics and handle gracefully. In the event this happens, we no longer get wedged because it will also cancel the context being used by other goroutines that are dependant on this work.