Use latest version of incubator/vindex
Updated to work with the latest API. This includes fixing method names,
and providing an Output Log.

This also removes the web form we used to serve, and instead brings up
an API consistent with the demo in the incubator project to allow the
client there to query the verifiable index.